// packages/shareds/focus-trap/src/types.ts
export type FocusableElement = HTMLElement | SVGElement
// ...
interface ContainerGroup {
  container: HTMLElement
  tabbableNodes: FocusableElement[]
  focusableNodes: FocusableElement[]
  posTabIndexesFound: boolean
  firstTabbableNode: FocusableElement | undefined
  lastTabbableNode: FocusableElement | undefined
  firstDomTabbableNode: FocusableElement | undefined
  lastDomTabbableNode: FocusableElement | undefined
  nextTabbableNode: (node: HTMLElement, forward?: boolean) => FocusableElement | undefined
}

export interface FindNextNodeOptions {
  target?: HTMLElement | null
  event?: FocusEvent | KeyboardEvent
  isBackward?: boolean
}

export interface FocusTrapState {
  containers: HTMLElement[]
  nodeFocusedBeforeActivation: HTMLElement | null
  mostRecentlyFocusedNode: HTMLElement | null
  active: boolean
  paused: boolean
  delayInitialFocusTimer: any
  containerGroups: ContainerGroup[]
  tabbableGroups: ContainerGroup[]
  // the most recent KeyboardEvent for the configured nav key (typically [SHIFT+]TAB), if any
  recentNavEvent: KeyboardEvent | undefined
}

export interface DeactivateOptions
  extends Pick<FocusTrapOptions, 'onDeactivate' | 'onPostDeactivate' | 'checkCanReturnFocus'> {
  returnFocus?: boolean | undefined
}

export type ActivateOptions = Pick<FocusTrapOptions, 'onActivate' | 'onPostActivate' | 'checkCanFocusTrap'>
export type PauseOptions = Pick<FocusTrapOptions, 'onPause' | 'onPostPause'>
export type UnpauseOptions = Pick<FocusTrapOptions, 'onUnpause' | 'onPostUnpause'>